Move revert to version to action buttons.
